+++ CHANGES.md
@@ -1,6 +1,18 @@
 CytoMeth - Changes
 =============
 
+## Version 0.9.8
+* README updated
+* QC extended
+* install packages extended by bioconductor packages
+* Dockerfile added
+
+## Version 0.9.7
+* README updated
+* FastQC QC report creation added
+* trimmomatic_MINLEN parameter added
+* seqtk added
+
 ## Version 0.9.6
 * GATK version 3.8.1
 * improved error handling and missing file checking